Refacto Skill Entity
