+++ src/Entity/BaseSkill.php
@@ -1,0 +1,70 @@
+<?php
+
+namespace AppBundle\Entity;
+
+use Doctrine\ORM\Mapping as ORM;
+use Gedmo\Mapping\Annotation as Gedmo;
+use Symfony\Component\Validator\Constraints as Assert;
+
+/**
+ * @ORM\MappedSuperclass
+ */
+abstract class BaseSkill
+{
+    /**
+     * @var int|null
+     *
+     * @ORM\Id
+     * @ORM\Column(type="integer")
+     * @ORM\GeneratedValue(strategy="IDENTITY")
+     */
+    private $id;
+
+    /**
+     * @var string
+     *
+     * @ORM\Column
+     *
+     * @Assert\NotBlank
+     * @Assert\Length(min=2, max=200)
+     */
+    private $name = '';
+
+    /**
+     * @var string|null
+     *
+     * @ORM\Column
+     * @Gedmo\Slug(fields={"name"}, unique=true)
+     */
+    private $slug;
+
+    public function __construct(?string $name = null)
+    {
+        $this->name = (string) $name;
+    }
+
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    public function setName(string $name): void
+    {
+        $this->name = $name;
+    }
+
+    public function getSlug(): ?string
+    {
+        return $this->slug;
+    }
+
+    public function __toString(): string
+    {
+        return $this->name;
+    }
+}
