+++ src/Entity/Skill.php
@@ -4,9 +4,7 @@ namespace AppBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
 use Doctrine\Common\Collections\ArrayCollection;
-use Gedmo\Mapping\Annotation as Gedmo;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
-use Symfony\Component\Validator\Constraints as Assert;
 
 /**
  * @ORM\Entity(repositoryClass="AppBundle\Repository\SkillRepository")
@@ -19,35 +17,8 @@ use Symfony\Component\Validator\Constraints as Assert;
  *
  * @UniqueEntity("name")
  */
-class Skill
+class Skill extends BaseSkill
 {
-    /**
-     * @var int|null
-     *
-     * @ORM\Id
-     * @ORM\Column(type="integer")
-     * @ORM\GeneratedValue(strategy="IDENTITY")
-     */
-    private $id;
-
-    /**
-     * @var string
-     *
-     * @ORM\Column
-     *
-     * @Assert\NotBlank
-     * @Assert\Length(min=2, max=200)
-     */
-    private $name = '';
-
-    /**
-     * @var string|null
-     *
-     * @ORM\Column
-     * @Gedmo\Slug(fields={"name"}, unique=true)
-     */
-    private $slug;
-
     /**
      * @ORM\ManyToMany(targetEntity="AppBundle\Entity\Summary", mappedBy="skills")
      *
@@ -64,7 +35,8 @@ class Skill
 
     public function __construct(?string $name = null)
     {
-        $this->name = (string) $name;
+        parent::__construct($name);
+
         $this->summaries = new ArrayCollection();
         $this->citizenInitiatives = new ArrayCollection();
     }
@@ -81,30 +53,5 @@ class Skill
         if (!$this->citizenInitiatives->contains($initiative)) {
             $this->citizenInitiatives->add($initiative);
         }
-    }
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getName(): string
-    {
-        return $this->name;
-    }
-
-    public function setName(string $name): void
-    {
-        $this->name = $name;
-    }
-
-    public function getSlug(): ?string
-    {
-        return $this->slug;
-    }
-
-    public function __toString(): string
-    {
-        return $this->name;
     }
 }
